added -Werrro g++ option to most projects to force errors, then, fixed all of the errors that were previously warnings. Most errors were related to signed/unsigned comparisons, so, many types were converted to unsigned if they could never be negative. Also fixed many errors of variables declared or declared and set, but then never used.
git-svn-id: http://svn.code.sf.net/p/mosat/code/trunk@526 becb1130-6c36-0410-badf-e7708b0ee230

+++ trainpos/ri_tplib.cpp
@@ -169,7 +169,7 @@ bool parseTimeString(const char* szWord)
 {
   unsigned currentState = 0;
   for (unsigned ss = 0; szWord[ss]; ss++)
-    currentState = g_TimeParserTransitionFunction[szWord[ss]][currentState];
+    currentState = g_TimeParserTransitionFunction[(unsigned)szWord[ss]][currentState];
   return 9 == currentState || 10 == currentState ? true : false;
 }
 
@@ -569,7 +569,7 @@ TPRC_ENUM ri_cfReader::loadConfigFile(const char* szPath)
 {
   m_Logger << INDENT << "Entering ri_cfReader::loadConfigFile...\n";
   m_Logger.increaseIndent();
-  long fSize;
+  unsigned long fSize;
   size_t nOfChars;
   TPRC_ENUM rCode;
   FILE* pConfigFile;
